pulling out credentials logic
